file processing generalized using ResultsFile.from_* methods
