+++ esofile_reader/__init__.py
@@ -1,6 +1,5 @@
-from esofile_reader.diff_file import DiffFile
 from esofile_reader.eso_file import EsoFile
 from esofile_reader.fetchers import get_results
 from esofile_reader.mini_classes import Variable
+from esofile_reader.results_file import ResultsFile
 from esofile_reader.storage.sql_storage import SQLFile
-from esofile_reader.totals_file import TotalsFile
