+++ esofile_reader/processing/interval_processor.py
