+++ tests/test_file_generation.py
@@ -1,6 +1,6 @@
 import unittest
 
-from esofile_reader import DiffFile, TotalsFile
+from esofile_reader import ResultsFile
 from tests import EF_ALL_INTERVALS, EF_ALL_INTERVALS_PEAKS
 
 
@@ -12,29 +12,29 @@ class TestFileGeneration(unittest.TestCase):
         self.assertTrue(EF_ALL_INTERVALS_PEAKS.complete)
 
     def test_eso_file_to_totals_file(self):
-        tf = TotalsFile(EF_ALL_INTERVALS)
+        tf = ResultsFile.from_totals(EF_ALL_INTERVALS)
         self.assertTrue(tf.complete)
 
     def test_diff_file(self):
-        df = DiffFile(EF_ALL_INTERVALS, EF_ALL_INTERVALS)
+        df = ResultsFile.from_diff(EF_ALL_INTERVALS, EF_ALL_INTERVALS)
         self.assertTrue(df.complete)
 
     def test_diff_file_to_totals_file(self):
-        df = DiffFile(EF_ALL_INTERVALS, EF_ALL_INTERVALS)
-        tf = TotalsFile(df)
+        df = ResultsFile.from_diff(EF_ALL_INTERVALS, EF_ALL_INTERVALS)
+        tf = ResultsFile.from_totals(df)
         self.assertTrue(tf.complete)
 
     def test_generate_totals_eso_file(self):
-        tf = EF_ALL_INTERVALS.generate_totals()
+        tf = ResultsFile.from_totals(EF_ALL_INTERVALS)
         self.assertTrue(tf.complete)
 
     def test_generate_diff_eso_file(self):
-        df = EF_ALL_INTERVALS.generate_diff(EF_ALL_INTERVALS_PEAKS)
+        df = ResultsFile.from_diff(EF_ALL_INTERVALS, EF_ALL_INTERVALS_PEAKS)
         self.assertTrue(df.complete)
 
     def test_generate_diff_totals_file(self):
-        tf = EF_ALL_INTERVALS.generate_totals()
-        df = tf.generate_diff(tf)
+        tf = ResultsFile.from_totals(EF_ALL_INTERVALS)
+        df = ResultsFile.from_diff(tf, tf)
         self.assertTrue(df.complete)
