+++ tests/test_interval_processor.py
@@ -2,8 +2,8 @@ import unittest
 from datetime import datetime
 
 from esofile_reader.mini_classes import IntervalTuple
-from esofile_reader.processor.interval_processor import *
-from esofile_reader.processor.interval_processor import _to_timestamp, _gen_dt
+from esofile_reader.processing.interval_processor import *
+from esofile_reader.processing.interval_processor import _to_timestamp, _gen_dt
 
 
 class TestIntervalProcessing(unittest.TestCase):
