+++ tests/test_parquet_storage.py
@@ -6,10 +6,9 @@ from pathlib import Path
 
 from pandas.testing import assert_frame_equal
 
-from esofile_reader import EsoFile
-from esofile_reader import TotalsFile
+from esofile_reader import EsoFile, ResultsFile
 from esofile_reader.data.pqt_data import ParquetFrame
-from esofile_reader.processor.monitor import DefaultMonitor
+from esofile_reader.processing.monitor import DefaultMonitor
 from esofile_reader.storage.pqt_storage import ParquetStorage, ParquetFile
 from tests import ROOT, EF1, EF_ALL_INTERVALS
 
@@ -19,7 +18,7 @@ logging.basicConfig(level=logging.INFO)
 class TestParquetStorage(unittest.TestCase):
     @classmethod
     def setUpClass(cls):
-        cls.tf = TotalsFile(EF1)
+        cls.tf = ResultsFile.from_totals(EF1)
 
     @classmethod
     def tearDownClass(cls):
@@ -42,13 +41,13 @@ class TestParquetStorage(unittest.TestCase):
     def test_02_store_file(self):
         self.storage.store_file(EF_ALL_INTERVALS)
         self.assertEqual("eplusout_all_intervals", self.storage.files[0].file_name)
-        self.assertEqual("EsoFile", self.storage.files[0].type_)
+        self.assertEqual("eso", self.storage.files[0].file_type)
 
     def test_03_store_totals_file(self):
-        tf = TotalsFile(EF_ALL_INTERVALS)
+        tf = ResultsFile.from_totals(EF_ALL_INTERVALS)
         self.storage.store_file(tf)
         self.assertEqual("eplusout_all_intervals - totals", self.storage.files[0].file_name)
-        self.assertEqual("TotalsFile", self.storage.files[0].type_)
+        self.assertEqual("totals", self.storage.files[0].file_type)
 
     def test_04_store_multiple_files(self):
         id1 = self.storage.store_file(EF_ALL_INTERVALS)
@@ -79,11 +78,11 @@ class TestParquetStorage(unittest.TestCase):
         pqf = ParquetFile.load_file("pqf" + ParquetFile.EXT, "")
         Path("pqf" + ParquetFile.EXT).unlink()
 
-        self.assertEqual(EF_ALL_INTERVALS.file_path, pqf.file_path)
+        self.assertEqual(str(EF_ALL_INTERVALS.file_path), str(pqf.file_path))
         self.assertEqual(EF_ALL_INTERVALS.file_name, pqf.file_name)
         self.assertEqual(EF_ALL_INTERVALS.file_created, pqf.file_created)
-        self.assertEqual(EF_ALL_INTERVALS.__class__.__name__, pqf.type_)
-        self.assertEqual(EF_ALL_INTERVALS.file_path, pqf.file_path)
+        self.assertEqual("eso", pqf.file_type)
+        self.assertEqual(str(EF_ALL_INTERVALS.file_path), str(pqf.file_path))
 
         for interval in EF_ALL_INTERVALS.table_names:
             assert_frame_equal(
@@ -103,30 +102,30 @@ class TestParquetStorage(unittest.TestCase):
         ParquetFrame.CHUNK_SIZE = 10
         self.storage.store_file(EF_ALL_INTERVALS)
         self.storage.store_file(EF1)
-        self.storage.store_file(TotalsFile(EF1))
+        self.storage.store_file(ResultsFile.from_totals(EF1))
         self.storage.save_as("", "pqs")
         self.assertTrue(Path("pqs" + ParquetStorage.EXT).exists())
 
     def test_09_load_storage(self):
         pqs = ParquetStorage.load_storage("pqs" + ParquetStorage.EXT)
 
-        self.assertEqual(EF_ALL_INTERVALS.file_path, pqs.files[0].file_path)
+        self.assertEqual(str(EF_ALL_INTERVALS.file_path), str(pqs.files[0].file_path))
         self.assertEqual(EF_ALL_INTERVALS.file_name, pqs.files[0].file_name)
         self.assertEqual(EF_ALL_INTERVALS.file_created, pqs.files[0].file_created)
-        self.assertEqual(EF_ALL_INTERVALS.__class__.__name__, pqs.files[0].type_)
-        self.assertEqual(EF_ALL_INTERVALS.file_path, pqs.files[0].file_path)
+        self.assertEqual(EF_ALL_INTERVALS.file_type, pqs.files[0].file_type)
+        self.assertEqual(str(EF_ALL_INTERVALS.file_path), pqs.files[0].file_path)
 
-        self.assertEqual(EF1.file_path, pqs.files[1].file_path)
+        self.assertEqual(str(EF1.file_path), str(pqs.files[1].file_path))
         self.assertEqual(EF1.file_name, pqs.files[1].file_name)
         self.assertEqual(EF1.file_created, pqs.files[1].file_created)
-        self.assertEqual(EF1.__class__.__name__, pqs.files[1].type_)
-        self.assertEqual(EF1.file_path, pqs.files[1].file_path)
+        self.assertEqual(EF1.file_type, pqs.files[1].file_type)
+        self.assertEqual(str(EF1.file_path), pqs.files[1].file_path)
 
-        self.assertEqual(self.tf.file_path, pqs.files[2].file_path)
+        self.assertEqual(str(self.tf.file_path), str(pqs.files[2].file_path))
         self.assertEqual(self.tf.file_name, pqs.files[2].file_name)
         self.assertEqual(self.tf.file_created, pqs.files[2].file_created)
-        self.assertEqual(self.tf.__class__.__name__, pqs.files[2].type_)
-        self.assertEqual(self.tf.file_path, pqs.files[2].file_path)
+        self.assertEqual("totals", pqs.files[2].file_type)
+        self.assertEqual(str(self.tf.file_path), str(pqs.files[2].file_path))
 
         for interval in EF_ALL_INTERVALS.table_names:
             assert_frame_equal(
@@ -169,7 +168,7 @@ class TestParquetStorage(unittest.TestCase):
         self.assertEqual(0, id_)
 
         monitor = DefaultMonitor("bar")
-        tf = TotalsFile(ef)
+        tf = ResultsFile.from_totals(ef)
         id_ = pqs.store_file(tf, monitor=monitor)
         self.assertEqual(1, id_)
 
@@ -216,7 +215,7 @@ class TestParquetStorage(unittest.TestCase):
                 data=EF_ALL_INTERVALS.data,
                 file_created=EF_ALL_INTERVALS.file_created,
                 search_tree=EF_ALL_INTERVALS.search_tree,
-                type_=EF_ALL_INTERVALS.__class__.__name__,
+                file_type=EF_ALL_INTERVALS.file_type,
                 pardir="",
                 name="foo",
                 monitor=None,
